+++ Changey.Tests/Commands/CompareCommandTests.cs
@@ -1,4 +1,5 @@
-﻿using System.Threading.Tasks;
+﻿using System;
+using System.Threading.Tasks;
 using Changey.Commands;
 using Changey.Options;
 using Changey.Services;
@@ -27,5 +28,27 @@ public class CompareCommandTests
 
 		// Assert
 		await generator.Received(1).Generate(path, baseUrl, compareTemplate, releaseTemplate);
+	}
+
+	[Fact]
+	public async Task ExecuteShouldLogWhenCreatorThrows()
+	{
+		// Arrange
+		var compareGenerator = Substitute.For<ICompareGenerator>();
+		compareGenerator.Generate(Arg.Any<string>(), Arg.Any<string>(), Arg.Any<string>(), Arg.Any<string>())
+			.Returns(Task.FromException<bool>(new Exception("test-exception")));
+
+		var option = new CompareOption("", "", "", "", false, false);
+
+		var logger = Substitute.For<ILogger>();
+		option.InjectLogger(logger);
+
+		var sut = new CompareCommand(option, compareGenerator);
+
+		// Act
+		await sut.Execute();
+
+		// Assert
+		logger.Received(1).Error(Arg.Any<string>(), Arg.Is<Exception>(e => e.Message.Contains("test-exception")));
 	}
 }