+++ Changey.Tests/TypeLoaderTests.cs
@@ -57,7 +57,7 @@ public class TypeLoaderTests
 	public void FindCommandShouldFindForInitOptions()
 	{
 		// Arrange
-		var option = new InitOption(true, true, string.Empty, false, false);
+		var option = new InitOption("", "", true, "", true, string.Empty, false, false);
 		var sut = new TypeLoader();
 
 		// Act