+++ Changey/Commands/CompareCommand.cs
@@ -1,4 +1,5 @@
-﻿using System.Threading.Tasks;
+﻿using System;
+using System.Threading.Tasks;
 using Changey.Options;
 using Changey.Services;
 
@@ -14,8 +15,15 @@ internal class CompareCommand : ICommand
 
 	public async Task Execute()
 	{
-		await _compareGenerator.Generate(_option.Path, _option.BaseUrl, _option.CompareTemplate,
-			_option.ReleaseTemplate);
+		try
+		{
+			await _compareGenerator.Generate(_option.Path, _option.BaseUrl, _option.CompareTemplate,
+				_option.ReleaseTemplate);
+		}
+		catch (Exception ex)
+		{
+			_option.Logger.Error("Failed to set up compare URLs", ex);
+		}
 	}
 
 	private readonly CompareOption _option;