+++ Changey/Commands/InitCommand.cs
@@ -7,13 +7,35 @@ namespace Changey.Commands;
 
 internal class InitCommand : ICommand
 {
-	public InitCommand(InitOption option, IChangeLogCreator changeLogCreator)
+	public InitCommand(InitOption option, IChangeLogCreator changeLogCreator, ICompareGenerator compareGenerator)
 	{
 		_option = option;
 		_changeLogCreator = changeLogCreator;
+		_compareGenerator = compareGenerator;
 	}
 
 	public async Task Execute()
+	{
+		await ExecuteInit();
+
+		if (!string.IsNullOrEmpty(_option.BaseUrl))
+			await ExecuteCompare();
+	}
+
+	private async Task ExecuteCompare()
+	{
+		try
+		{
+			await _compareGenerator.Generate(_option.Path, _option.BaseUrl, _option.CompareTemplate,
+				_option.ReleaseTemplate);
+		}
+		catch (Exception ex)
+		{
+			_option.Logger.Error("Failed to set up compare URLs", ex);
+		}
+	}
+
+	private async Task ExecuteInit()
 	{
 		try
 		{
@@ -29,4 +51,5 @@ internal class InitCommand : ICommand
 
 	private readonly InitOption _option;
 	private readonly IChangeLogCreator _changeLogCreator;
+	private readonly ICompareGenerator _compareGenerator;
 }