+++ changelog.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Allow init command to set up compare URLs instead of needing two commands
+
 ## [0.4.1] - 2022-05-29
 ### Added
 - Add support for compare URLs to view changes between versions (#28)